Adding Ajax to cart
Moved line_item to a template and updated cart to use it
Moved cart to a template and modified it to used a cart passed as a parameter, not from the session.
Modified cart.show to use the template so it still works when JS is disabled.
Updated the application.html.erb to render the cart in the sidebar.
Modified the StoreController to set/setup the cart using the cart concern.
Redirecting line_item.create success to store_index_url to refresh the page whenever an item is added to cart